backend/requirements/common.txt
Django==3.2.18
djangorestframework
django-cors-headers
pillow